task controller and task router created

+++ src/controllers/user.controller.ts
@@ -108,6 +108,7 @@ export default class UserController {
         error: false,
         data: {
           user: {
+            user_id: response.rows[0].user_id,
             username,
             user_email,
           },
@@ -136,7 +137,7 @@ export default class UserController {
     }
     try {
       const response: QueryResult = await pool.query(
-        "SELECT user_id, username, user_email, user_password FROM users WHERE user_email = $1",
+        "SELECT * FROM users WHERE user_email = $1",
         [user_email]
       );
       if (response.rowCount <= 0) {
@@ -168,6 +169,7 @@ export default class UserController {
         error: false,
         data: {
           user: {
+            user_id: response.rows[0].user_id,
             username: response.rows[0].username,
             user_email: response.rows[0].user_email,
           },
@@ -236,12 +238,12 @@ export default class UserController {
         const hash: string = await hashPassword(body.user_password);
         body.user_password = hash;
         response = await pool.query(
-          "UPDATE users SET username = $1, user_email = $2, user_password = $3 WHERE user_id = $4 RETURNING username, user_email",
+          "UPDATE users SET username = $1, user_email = $2, user_password = $3 WHERE user_id = $4 RETURNING user_id username, user_email",
           [body.username, body.user_email, body.user_password, user_id]
         );
       } else {
         response = await pool.query(
-          "UPDATE users SET username = $1, user_email = $2 WHERE user_id = $3 RETURNING username, user_email",
+          "UPDATE users SET username = $1, user_email = $2 WHERE user_id = $3 RETURNING user_id, username, user_email",
           [body.username, body.user_email, user_id]
         );
       }
@@ -274,7 +276,7 @@ export default class UserController {
 
     try {
       const response: QueryResult = await pool.query(
-        "DELETE FROM users WHERE user_id = $1 RETURNING username, user_email",
+        "DELETE FROM users WHERE user_id = $1 RETURNING user_id, username, user_email",
         [user_id]
       );
       if (response.rowCount <= 0) {
@@ -294,7 +296,7 @@ export default class UserController {
   public static async getUsers(req: Request, res: Response): Promise<Response> {
     try {
       const response: QueryResult = await pool.query(
-        "SELECT username, user_email FROM users"
+        "SELECT user_id, username, user_email FROM users"
       );
       return res.status(200).send(response.rows);
     } catch (err) {
@@ -309,7 +311,7 @@ export default class UserController {
     const { params } = req;
     try {
       const response: QueryResult = await pool.query(
-        "SELECT username, user_email FROM users WHERE user_id = $1",
+        "SELECT user_id, username, user_email FROM users WHERE user_id = $1",
         [params.id]
       );
       if (response.rowCount <= 0) {
